Adição do HATEOAS na atividade 4

+++ ATVIV/automanager/src/main/java/com/autobots/automanager/controles/ControleServico.java
@@ -6,60 +6,58 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.*;
 
 import com.autobots.automanager.entidades.Servico;
 import com.autobots.automanager.repositorios.RepositorioServico;
 
-
 @RestController
 @RequestMapping("/servicos")
-public class ControleServico {
+public class ControleServico extends ControleHATEOAS<Servico> {
 
     @Autowired
     private RepositorioServico repositorio;
 
-    @GetMapping
-    @PreAuthorize("hasAnyRole('ADMIN', 'GERENTE', 'VENDEDOR')")
-    public ResponseEntity<List<Servico>> obterServicos() {
-        List<Servico> servicos = repositorio.findAll();
-        return new ResponseEntity<>(servicos, HttpStatus.OK);
-    }
-
-    @GetMapping("/{id}")
-    @PreAuthorize("hasAnyRole('ADMIN', 'GERENTE', 'VENDEDOR')")
-    public ResponseEntity<Servico> obterServico(@PathVariable Long id) {
-        Servico servico = repositorio.findById(id).orElse(null);
-        return servico != null 
-            ? new ResponseEntity<>(servico, HttpStatus.OK)
-            : new ResponseEntity<>(HttpStatus.NOT_FOUND);
-    }
-
-    @PostMapping
     @PreAuthorize("hasAnyRole('ADMIN', 'GERENTE')")
+    @PostMapping("/cadastrar")
     public ResponseEntity<?> cadastrarServico(@RequestBody Servico servico) {
         repositorio.save(servico);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
-    @PutMapping
-    @PreAuthorize("hasAnyRole('ADMIN', 'GERENTE')")
-    public ResponseEntity<?> atualizarServico(@RequestBody Servico servico) {
-        repositorio.save(servico);
-        return new ResponseEntity<>(HttpStatus.OK);
+    @PreAuthorize("hasAnyRole('ADMIN', 'GERENTE', 'VENDEDOR')")
+    @GetMapping("/listar")
+    public ResponseEntity<List<Servico>> listarServicos() {
+        List<Servico> servicos = repositorio.findAll();
+        return configurarResposta(servicos, this.getClass());
     }
 
-    @DeleteMapping("/{id}")
-    @PreAuthorize("hasRole('ADMIN')")
-    public ResponseEntity<?> excluirServico(@PathVariable Long id) {
-        repositorio.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+    @PreAuthorize("hasAnyRole('ADMIN', 'GERENTE')")
+    @GetMapping("/buscar/{id}")
+    public ResponseEntity<Servico> buscarServico(@PathVariable Long id) {
+        Servico servico = repositorio.findById(id).orElse(null);
+        return configurarResposta(servico, this.getClass());
+    }
+    @PreAuthorize("hasAnyRole('ADMIN', 'GERENTE')")
+    @DeleteMapping("/deletar/{id}")
+    public ResponseEntity<?> deletarServico(@PathVariable Long id) {
+        if (repositorio.existsById(id)) {
+            repositorio.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+    @PreAuthorize("hasAnyRole('ADMIN', 'GERENTE')")
+    @PutMapping("/atualizar/{id}")
+    public ResponseEntity<?> atualizarServico(@PathVariable Long id, @RequestBody Servico servicoAtualizado) {
+        if (repositorio.existsById(id)) {
+            Servico servicoExistente = repositorio.findById(id).orElse(null);
+            if (servicoExistente != null) {
+                servicoAtualizado.setId(servicoExistente.getId());
+                repositorio.save(servicoAtualizado);
+                return new ResponseEntity<>(HttpStatus.OK);
+            }
+        }
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 }
